requirements.txt
# Django Book Management System - Requirements
# Production Ready for cPanel (scholarquest.tech)

# Core
Django>=4.2.0,<5.0.0
psycopg2-binary>=2.9.5
Pillow>=10.0.0
python-dotenv>=1.0.0

# Translation
django-rosetta>=0.10.0
django-modeltranslation>=0.20.0
translatebot-django>=1.3.0

# Data & Analytics
pandas>=2.0.0
plotly>=5.0.0
openpyxl>=3.1.0

# Utilities
django-extensions>=3.2.0
markdown>=3.5.0

# Data migration (optional)
PyMySQL>=1.0.2
